fix: change Julian Day to calendar date conversion
The formula for converting a Julian Day to a calendar date did not
work correctly in all cases. The formula was replaced by the one
described by Jean Meeus in Astronomical Algorithms,
First English Edition, page 63.

A unit test for the failing case was added.

Thanks to Renaud S. for discovering the error.